2nd Private Leaderboard - MaD Lab team/Code/Training notebooks/training/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm == 0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy == 1.6.7
decorator == 5.1.1
docopt==0.6.2
executing == 2.0.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.65.5
huggingface-hub==0.24.6
idna==3.7
imageio==2.35.1
importlib_metadata==8.3.0
importlib_resources==6.4.3
ipykernel==6.29.5
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpickle==3.2.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
nest_asyncio==1.6.0
networkx==3.2.1
numpy==2.0.1
opencv-python==4.10.0.84
openslide-python==1.3.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==24.0.1
regex==2024.7.24
requests==2.32.3
sacred==0.8.5
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
smmap==5.0.1
spams-bin==2.6.10
stack-data==0.6.2
stainlib @ git+https://github.com/sebastianffx/stainlib.git@cf0b60352404f60d2a797d04b343bad22013ce10
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.8.10
tokenizers==0.19.1
torch==2.2.2
torchmetrics==1.4.1
torchstain @ git+https://github.com/EIDOSLAB/torchstain.git@3a3eb2c8af2662b4a001cf134f9a1bd24e0bdba7
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.0
